pyproject.toml
[project]
name = "ai-teacher"
version = "0.1.0"
description = "AI-powered platform for studies"
authors = [
    {name = "FernandoAurelius, CarlosFAlmada"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "markdown (>=3.8.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "google-genai (>=1.31.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
